Reorganized Charts dir
