+++ front/src/Charts/Bool/Bool-test.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import { expect } from 'chai'
 import { shallow, mount } from 'enzyme'
-import Bool from '../src/charts/Bool'
+import Bool from './Bool'
 
 const arraySite = require('./resources/array-sites')
