+++ front/src/Charts/Bool/Bool.js
