+++ front/src/Charts/Categories/Categories.js
