+++ front/src/Content/Content.js
@@ -1,7 +1,7 @@
 import React, { Component } from 'react'
 import Site from '../Site/Site'
 import SearchBar from '../SearchBar/SearchBar'
-import Charts from '../Charts'
+import Charts from '../Charts/Charts'
 import './Content.css'
 
 function getParameterByName(name, url) {
